Update from cookiecutter

+++ tests/test_bake_project.py
@@ -58,7 +58,8 @@ def bake_in_temp_dir(cookies, *args, **kwargs):
     try:
         yield result
     finally:
-        rmtree(str(result.project_path))
+        if not ("--keep-baked-projects" in sys.argv):
+            rmtree(str(result.project_path))
 
 
 def run_inside_dir(command, dirpath):
